doctor_patient/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
diskcache==5.6.3
    # via pyautogen
distro==1.9.0
    # via openai
docker==7.1.0
    # via pyautogen
flaml==2.1.2
    # via pyautogen
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
numpy==1.26.4
    # via
    #   flaml
    #   pyautogen
openai==1.35.5
    # via pyautogen
packaging==24.1
    # via pyautogen
pyautogen==0.2.31
    # via -r requirements.in
pydantic==2.7.4
    # via
    #   openai
    #   pyautogen
pydantic-core==2.18.4
    # via pydantic
python-dotenv==1.0.1
    # via pyautogen
regex==2024.5.15
    # via tiktoken
requests==2.32.3
    # via
    #   docker
    #   tiktoken
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
termcolor==2.4.0
    # via pyautogen
tiktoken==0.7.0
    # via pyautogen
tqdm==4.66.4
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.2
    # via
    #   docker
    #   requests
